refactor(lsp): split import candidate collection from ranking and rendering (#9420)
Split crates/tsz-lsp/src/project/imports.rs (1697 lines) into four
responsibility-scoped modules under imports/:

- mod.rs (231L): entry-point methods and orchestration
- collection.rs (869L): per-file AST export-match traversal
- filtering.rs (391L): exclusion predicates, bare-specifier logic, BareSpecifierSourceCache
- position.rs (295L): cursor/identifier resolution and import-target navigation

Behavior-preserving: no resolution, ordering, or LSP output changes.
